Correctly implement and use Actor base class
Our MgActor class was not actually extending the Foundry Actor class.
The data we were trying to wrap is actually the "actor.data.data" field,
so our definition of the actor subclass was incorrect.

Create an actual MgActor class the extends Actor and overrides the
prepare*Data() handlers methods.

Our old Actor class is now MgActorData, which we now convert the
actor.data.data field to in prepareDerivedData().

This way, we setup the hooks for the actor correctly, and now have
a class the wraps the data field with a strictly typed class.

character-sheet.ts is functionally the same, but has been updated to
use the new class names.